+++ gluecodium/src/test/resources/smoke/nullable/output/dart/ffi/ffi_smoke_Nullable.cpp
@@ -19,7 +19,7 @@
 extern "C" {
 #endif
 FfiOpaqueHandle
-library_smoke_Nullable_methodWithString__String(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
+library_smoke_Nullable_methodWithString__String_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
     return gluecodium::ffi::Conversion<gluecodium::optional<std::string>>::toFfi(
         (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).method_with_string(
@@ -28,7 +28,7 @@ library_smoke_Nullable_methodWithString__String(FfiOpaqueHandle _self, int32_t _
     );
 }
 FfiOpaqueHandle
-library_smoke_Nullable_methodWithBoolean__Boolean(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
+library_smoke_Nullable_methodWithBoolean__Boolean_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
     return gluecodium::ffi::Conversion<gluecodium::optional<bool>>::toFfi(
         (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).method_with_boolean(
@@ -37,7 +37,7 @@ library_smoke_Nullable_methodWithBoolean__Boolean(FfiOpaqueHandle _self, int32_t
     );
 }
 FfiOpaqueHandle
-library_smoke_Nullable_methodWithDouble__Double(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
+library_smoke_Nullable_methodWithDouble__Double_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
     return gluecodium::ffi::Conversion<gluecodium::optional<double>>::toFfi(
         (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).method_with_double(
@@ -46,7 +46,7 @@ library_smoke_Nullable_methodWithDouble__Double(FfiOpaqueHandle _self, int32_t _
     );
 }
 FfiOpaqueHandle
-library_smoke_Nullable_methodWithInt__Long(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
+library_smoke_Nullable_methodWithInt__Long_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
     return gluecodium::ffi::Conversion<gluecodium::optional<int64_t>>::toFfi(
         (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).method_with_int(
@@ -55,7 +55,7 @@ library_smoke_Nullable_methodWithInt__Long(FfiOpaqueHandle _self, int32_t _isola
     );
 }
 FfiOpaqueHandle
-library_smoke_Nullable_methodWithSomeStruct__SomeStruct(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
+library_smoke_Nullable_methodWithSomeStruct__SomeStruct_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
     return gluecodium::ffi::Conversion<gluecodium::optional<smoke::Nullable::SomeStruct>>::toFfi(
         (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).method_with_some_struct(
@@ -64,7 +64,7 @@ library_smoke_Nullable_methodWithSomeStruct__SomeStruct(FfiOpaqueHandle _self, i
     );
 }
 FfiOpaqueHandle
-library_smoke_Nullable_methodWithSomeEnum__SomeEnum(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
+library_smoke_Nullable_methodWithSomeEnum__SomeEnum_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
     return gluecodium::ffi::Conversion<gluecodium::optional<smoke::Nullable::SomeEnum>>::toFfi(
         (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).method_with_some_enum(
@@ -73,7 +73,7 @@ library_smoke_Nullable_methodWithSomeEnum__SomeEnum(FfiOpaqueHandle _self, int32
     );
 }
 FfiOpaqueHandle
-library_smoke_Nullable_methodWithSomeArray__ListOf_String(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
+library_smoke_Nullable_methodWithSomeArray__ListOf_String_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
     return gluecodium::ffi::Conversion<gluecodium::optional<std::vector<std::string>>>::toFfi(
         (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).method_with_some_array(
@@ -82,7 +82,7 @@ library_smoke_Nullable_methodWithSomeArray__ListOf_String(FfiOpaqueHandle _self,
     );
 }
 FfiOpaqueHandle
-library_smoke_Nullable_methodWithInlineArray__ListOf_String(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
+library_smoke_Nullable_methodWithInlineArray__ListOf_String_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
     return gluecodium::ffi::Conversion<gluecodium::optional<std::vector<std::string>>>::toFfi(
         (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).method_with_inline_array(
@@ -91,7 +91,7 @@ library_smoke_Nullable_methodWithInlineArray__ListOf_String(FfiOpaqueHandle _sel
     );
 }
 FfiOpaqueHandle
-library_smoke_Nullable_methodWithSomeMap__MapOf_Long_to_String(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
+library_smoke_Nullable_methodWithSomeMap__MapOf_Long_to_String_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
     return gluecodium::ffi::Conversion<gluecodium::optional<std::unordered_map<int64_t, std::string>>>::toFfi(
         (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).method_with_some_map(
@@ -100,7 +100,7 @@ library_smoke_Nullable_methodWithSomeMap__MapOf_Long_to_String(FfiOpaqueHandle _
     );
 }
 FfiOpaqueHandle
-library_smoke_Nullable_methodWithInstance__SomeInterface(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
+library_smoke_Nullable_methodWithInstance__SomeInterface_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle input) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
     return gluecodium::ffi::Conversion<std::shared_ptr<smoke::SomeInterface>>::toFfi(
         (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).method_with_instance(
@@ -116,7 +116,7 @@ library_smoke_Nullable_stringProperty_get(FfiOpaqueHandle _self, int32_t _isolat
     );
 }
 void
-library_smoke_Nullable_stringProperty_set__String(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
+library_smoke_Nullable_stringProperty_set__String_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
             (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).set_string_property(
             gluecodium::ffi::Conversion<gluecodium::optional<std::string>>::toCpp(value)
@@ -130,7 +130,7 @@ library_smoke_Nullable_isBoolProperty_get(FfiOpaqueHandle _self, int32_t _isolat
     );
 }
 void
-library_smoke_Nullable_isBoolProperty_set__Boolean(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
+library_smoke_Nullable_isBoolProperty_set__Boolean_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
             (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).set_bool_property(
             gluecodium::ffi::Conversion<gluecodium::optional<bool>>::toCpp(value)
@@ -144,7 +144,7 @@ library_smoke_Nullable_doubleProperty_get(FfiOpaqueHandle _self, int32_t _isolat
     );
 }
 void
-library_smoke_Nullable_doubleProperty_set__Double(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
+library_smoke_Nullable_doubleProperty_set__Double_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
             (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).set_double_property(
             gluecodium::ffi::Conversion<gluecodium::optional<double>>::toCpp(value)
@@ -158,7 +158,7 @@ library_smoke_Nullable_intProperty_get(FfiOpaqueHandle _self, int32_t _isolate_i
     );
 }
 void
-library_smoke_Nullable_intProperty_set__Long(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
+library_smoke_Nullable_intProperty_set__Long_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
             (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).set_int_property(
             gluecodium::ffi::Conversion<gluecodium::optional<int64_t>>::toCpp(value)
@@ -172,7 +172,7 @@ library_smoke_Nullable_structProperty_get(FfiOpaqueHandle _self, int32_t _isolat
     );
 }
 void
-library_smoke_Nullable_structProperty_set__SomeStruct(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
+library_smoke_Nullable_structProperty_set__SomeStruct_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
             (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).set_struct_property(
             gluecodium::ffi::Conversion<gluecodium::optional<smoke::Nullable::SomeStruct>>::toCpp(value)
@@ -186,7 +186,7 @@ library_smoke_Nullable_enumProperty_get(FfiOpaqueHandle _self, int32_t _isolate_
     );
 }
 void
-library_smoke_Nullable_enumProperty_set__SomeEnum(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
+library_smoke_Nullable_enumProperty_set__SomeEnum_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
             (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).set_enum_property(
             gluecodium::ffi::Conversion<gluecodium::optional<smoke::Nullable::SomeEnum>>::toCpp(value)
@@ -200,7 +200,7 @@ library_smoke_Nullable_arrayProperty_get(FfiOpaqueHandle _self, int32_t _isolate
     );
 }
 void
-library_smoke_Nullable_arrayProperty_set__ListOf_String(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
+library_smoke_Nullable_arrayProperty_set__ListOf_String_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
             (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).set_array_property(
             gluecodium::ffi::Conversion<gluecodium::optional<std::vector<std::string>>>::toCpp(value)
@@ -214,7 +214,7 @@ library_smoke_Nullable_inlineArrayProperty_get(FfiOpaqueHandle _self, int32_t _i
     );
 }
 void
-library_smoke_Nullable_inlineArrayProperty_set__ListOf_String(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
+library_smoke_Nullable_inlineArrayProperty_set__ListOf_String_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
             (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).set_inline_array_property(
             gluecodium::ffi::Conversion<gluecodium::optional<std::vector<std::string>>>::toCpp(value)
@@ -228,7 +228,7 @@ library_smoke_Nullable_mapProperty_get(FfiOpaqueHandle _self, int32_t _isolate_i
     );
 }
 void
-library_smoke_Nullable_mapProperty_set__MapOf_Long_to_String(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
+library_smoke_Nullable_mapProperty_set__MapOf_Long_to_String_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
             (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).set_map_property(
             gluecodium::ffi::Conversion<gluecodium::optional<std::unordered_map<int64_t, std::string>>>::toCpp(value)
@@ -242,7 +242,7 @@ library_smoke_Nullable_instanceProperty_get(FfiOpaqueHandle _self, int32_t _isol
     );
 }
 void
-library_smoke_Nullable_instanceProperty_set__SomeInterface(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
+library_smoke_Nullable_instanceProperty_set__SomeInterface_(FfiOpaqueHandle _self, int32_t _isolate_id, FfiOpaqueHandle value) {
     gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
             (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::Nullable>>::toCpp(_self)).set_instance_property(
             gluecodium::ffi::Conversion<std::shared_ptr<smoke::SomeInterface>>::toCpp(value)