+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   * Removed suppression of "dart analyze" warning about deprecated element usage. This should be suppressed in a
     library-level analysis_options.yaml instead.
   * Fixed "dart analyze" warnings about unused imports.
+  * Fixed validation of function overloads for C++, Java, and Swift generators; and related issues.
 
 ## 9.3.5
 Release date: 2021-07-28